linera_execution

Enum ExecutionError

pub enum ExecutionError {

    ViewError(ViewError),
    ArithmeticError(ArithmeticError),
    UserError(String),
    WasmError(WasmExecutionError),
    EvmError(EvmExecutionError),
    DecompressionError(DecompressionError),
    InvalidPromise,
    ReentrantCall(ApplicationId),
    CrossApplicationCallInFinalize {
        caller_id: Box<ApplicationId>,
        callee_id: Box<ApplicationId>,
    },
    ApplicationBytecodeNotFound(Box<ApplicationDescription>),
    UnsupportedDynamicApplicationLoad(Box<ApplicationId>),
    ExcessiveRead,
    ExcessiveWrite,
    MaximumFuelExceeded(VmRuntime),
    MaximumServiceOracleExecutionTimeExceeded,
    ServiceOracleResponseTooLarge,
    BlockTooLarge,
    HttpResponseSizeLimitExceeded {
        limit: u64,
        size: u64,
    },
    MissingRuntimeResponse,
    UnauthorizedApplication(ApplicationId),
    ReqwestError(Error),
    IoError(Error),
    UnexpectedOracleResponse,
    JsonError(Error),
    BcsError(Error),
    OracleResponseMismatch,
    ServiceOracleQueryOperations(Vec<Operation>),
    AssertBefore {
        timestamp: Timestamp,
        local_time: Timestamp,
    },
    StreamNameTooLong,
    BlobTooLarge,
    BytecodeTooLarge,
    UnauthorizedHttpRequest(Url),
    InvalidUrlForHttpRequest(ParseError),
    Thread(Error),
    InactiveChain(ChainId),
    BlobsNotFound(Vec<BlobId>),
    EventsNotFound(Vec<EventId>),
    InvalidHeaderName(InvalidHeaderName),
    InvalidHeaderValue(InvalidHeaderValue),
    NoNetworkDescriptionFound,
    InvalidEpoch {
        chain_id: ChainId,
        epoch: Epoch,
    },
    IncorrectTransferAmount,
    UnauthenticatedTransferOwner,
    InsufficientBalance {
        balance: Amount,
        account: AccountOwner,
    },
    FeesExceedFunding {
        fees: Amount,
        balance: Amount,
    },
    IncorrectClaimAmount,
    UnauthenticatedClaimOwner,
    AdminOperationOnNonAdminChain,
    InvalidCommitteeEpoch {
        expected: Epoch,
        provided: Epoch,
    },
    InvalidCommitteeRemoval,
    MissingOracleResponse,
    UnprocessedStreams,
    InternalError(&'static str),
    OutdatedUpdateStreams,

}

A type for errors happening during execution.

Variants

ViewError(ViewError)

ArithmeticError(ArithmeticError)

UserError(String)

WasmError(WasmExecutionError)

EvmError(EvmExecutionError)

DecompressionError(DecompressionError)

InvalidPromise

ReentrantCall(ApplicationId)

CrossApplicationCallInFinalize

Fields

caller_id: Box<ApplicationId>

callee_id: Box<ApplicationId>

ApplicationBytecodeNotFound(Box<ApplicationDescription>)

UnsupportedDynamicApplicationLoad(Box<ApplicationId>)

ExcessiveRead

ExcessiveWrite

MaximumFuelExceeded(VmRuntime)

MaximumServiceOracleExecutionTimeExceeded

ServiceOracleResponseTooLarge

BlockTooLarge

HttpResponseSizeLimitExceeded

Fields

limit: u64

size: u64

MissingRuntimeResponse

UnauthorizedApplication(ApplicationId)

ReqwestError(Error)

IoError(Error)

UnexpectedOracleResponse

JsonError(Error)

BcsError(Error)

OracleResponseMismatch

ServiceOracleQueryOperations(Vec<Operation>)

AssertBefore

Fields

timestamp: Timestamp

local_time: Timestamp

StreamNameTooLong

BlobTooLarge

BytecodeTooLarge

UnauthorizedHttpRequest(Url)

InvalidUrlForHttpRequest(ParseError)

Thread(Error)

InactiveChain(ChainId)

BlobsNotFound(Vec<BlobId>)

EventsNotFound(Vec<EventId>)

InvalidHeaderName(InvalidHeaderName)

InvalidHeaderValue(InvalidHeaderValue)

NoNetworkDescriptionFound

InvalidEpoch

Fields

chain_id: ChainId

epoch: Epoch

IncorrectTransferAmount

UnauthenticatedTransferOwner

InsufficientBalance

Fields

balance: Amount

account: AccountOwner

FeesExceedFunding

Fields

fees: Amount

balance: Amount

IncorrectClaimAmount

UnauthenticatedClaimOwner

AdminOperationOnNonAdminChain

InvalidCommitteeEpoch

Fields

expected: Epoch

provided: Epoch

InvalidCommitteeRemoval

MissingOracleResponse

UnprocessedStreams

InternalError(&'static str)

OutdatedUpdateStreams

Implementations

impl ExecutionError

pub fn is_local(&self) -> bool

Returns whether this error is caused by an issue in the local node.

Returns false whenever the error could be caused by a bad message from a peer.