Struct linera_execution::system::SystemExecutionStateView

pub struct SystemExecutionStateView<C> {
    pub description: RegisterView<C, Option<ChainDescription>>,
    pub epoch: RegisterView<C, Option<Epoch>>,
    pub admin_id: RegisterView<C, Option<ChainId>>,
    pub subscriptions: SetView<C, ChannelSubscription>,
    pub committees: RegisterView<C, BTreeMap<Epoch, Committee>>,
    pub ownership: RegisterView<C, ChainOwnership>,
    pub balance: RegisterView<C, Balance>,
    pub balances: MapView<C, Owner, Balance>,
    pub timestamp: RegisterView<C, Timestamp>,
    pub registry: ApplicationRegistryView<C>,
}

A view accessing the execution state of the system of a chain.

Fields
description: RegisterView<C, Option<ChainDescription>>

How the chain was created. May be unknown for inactive chains.

epoch: RegisterView<C, Option<Epoch>>

The number identifying the current configuration.

admin_id: RegisterView<C, Option<ChainId>>

The admin of the chain.

subscriptions: SetView<C, ChannelSubscription>

Track the channels that we have subscribed to.

committees: RegisterView<C, BTreeMap<Epoch, Committee>>

The committees that we trust, indexed by epoch number. Not using a MapView because the set active of committees is supposed to be small. Plus, currently, we would create the BTreeMap anyway in various places (e.g. the OpenChain operation).

ownership: RegisterView<C, ChainOwnership>

Ownership of the chain.

balance: RegisterView<C, Balance>

Balance of the chain (unattributed).

balances: MapView<C, Owner, Balance>

Balances attributed to a given owner.

timestamp: RegisterView<C, Timestamp>

The timestamp of the most recent block.

registry: ApplicationRegistryView<C>

Track the locations of known bytecodes as well as the descriptions of known applications.

Implementations

impl<C> SystemExecutionStateView<C>where C: Context + Clone + Send + Sync + 'static, ViewError: From<C::Error>,

pub fn is_active(&self) -> bool

Invariant for the states of active chains.

pub async fn execute_operation( &mut self, context: &OperationContext, operation: &SystemOperation ) -> Result<(RawExecutionResult<SystemEffect>, Option<(UserApplicationId, Vec<u8>)>), SystemExecutionError>

Executes the sender’s side of an operation and returns a list of actions to be taken.

pub async fn execute_effect( &mut self, context: &EffectContext, effect: &SystemEffect ) -> Result<RawExecutionResult<SystemEffect>, SystemExecutionError>

Executes the recipient’s side of an operation, aka a “remote effect”.

Effects should not return an error unless it is a temporary failure (e.g. storage) or a global system failure. An error will fail the entire cross-chain request, allowing it to be retried later.

If execution is impossible for a deterministic reason (e.g. insufficient funds), effects should fail silently and be skipped (similar to a transaction in traditional blockchains).

pub fn open_chain( &mut self, effect_id: EffectId, public_key: PublicKey, epoch: Epoch, committees: BTreeMap<Epoch, Committee>, admin_id: ChainId, timestamp: Timestamp )

Initializes the system application state on a newly opened chain.

pub async fn query_application( &mut self, context: &QueryContext, _query: &SystemQuery ) -> Result<SystemResponse, SystemExecutionError>