linera_execution

Module test_utils

Helpers for writing tests that exercise the execution layer.

Modules

solidity

Code for compiling solidity smart contracts for testing purposes.

Structs

MockApplication

A mocked implementation of a user application.

MockApplicationInstance

A mocked implementation of a user application instance.

SystemExecutionState

A system execution state, not represented as a view but as a simple struct.

Enums

ExpectedCall

An expected call to a MockApplicationInstance.

Traits

RegisterMockApplication

Registration of MockApplications to use in tests.

Functions

blob_oracle_responses

Creates a vector of OracleResponses for the supplied BlobIds.

create_dummy_message_context

Creates a dummy MessageContext to use in tests.

create_dummy_operation_context

Creates a dummy OperationContext to use in tests.

create_dummy_query_context

Creates a dummy QueryContext to use in tests.

create_dummy_user_application_description

Creates a dummy ApplicationDescription for use in tests.

create_dummy_user_application_registrations

Creates the given number of dummy user application registrations for use in tests.

dummy_chain_description

Creates a dummy chain description with a deterministic single owner and the maximum balance.

dummy_chain_description_with_owner

Creates a dummy chain description owned by the given owner, with the maximum balance.

dummy_chain_description_with_ownership_and_balance

Creates a dummy root chain description with the given index, ownership, and balance.

dummy_committee

Creates a dummy committee with a single test validator.

dummy_committees

Creates a dummy committee map containing a single committee at epoch zero.

test_accounts_strategy

Creates a Strategy for creating a BTreeMap of AccountOwners with an initial non-zero Amount of tokens.