Trait linera_execution::BaseRuntime

pub trait BaseRuntime: Send + Sync {
    fn chain_id(&self) -> ChainId;
    fn application_id(&self) -> UserApplicationId;
    fn application_parameters(&self) -> Vec<u8>;
    fn read_system_balance(&self) -> Balance;
    fn read_system_timestamp(&self) -> Timestamp;
    fn try_read_my_state<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>, ExecutionError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn lock_view_user_state<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(), ExecutionError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn unlock_view_user_state<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(), ExecutionError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_key_bytes<'life0, 'async_trait>(
        &'life0 self,
        key: Vec<u8>
    ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>, ExecutionError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn find_keys_by_prefix<'life0, 'async_trait>(
        &'life0 self,
        key_prefix: Vec<u8>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Vec<u8>>, ExecutionError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn find_key_values_by_prefix<'life0, 'async_trait>(
        &'life0 self,
        key_prefix: Vec<u8>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<(Vec<u8>, Vec<u8>)>, ExecutionError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Required Methods

fn chain_id(&self) -> ChainId

The current chain id.

fn application_id(&self) -> UserApplicationId

The current application id.

fn application_parameters(&self) -> Vec<u8>

The current application parameters.

fn read_system_balance(&self) -> Balance

Reads the system balance.

fn read_system_timestamp(&self) -> Timestamp

Reads the system timestamp.

fn try_read_my_state<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<u8>, ExecutionError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Reads the application state.

fn lock_view_user_state<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<(), ExecutionError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Locks the view user state and prevents further reading/loading

fn unlock_view_user_state<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<(), ExecutionError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Unlocks the view user state and allows reading/loading again

fn read_key_bytes<'life0, 'async_trait>(
&'life0 self,
key: Vec<u8>
) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>, ExecutionError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Reads the key from the KV store

fn find_keys_by_prefix<'life0, 'async_trait>(
&'life0 self,
key_prefix: Vec<u8>
) -> Pin<Box<dyn Future<Output = Result<Vec<Vec<u8>>, ExecutionError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Reads the data from the keys having a specific prefix.

fn find_key_values_by_prefix<'life0, 'async_trait>(
&'life0 self,
key_prefix: Vec<u8>
) -> Pin<Box<dyn Future<Output = Result<Vec<(Vec<u8>, Vec<u8>)>, ExecutionError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Reads the data from the key/values having a specific prefix.