linera_execution

Module wasm_test

This assumes that the current directory is one of the crates.

Functions

get_example_bytecode_paths

Returns the contract and service bytecode file paths for the named example application.