Crate linera_execution

This module manages the execution of the system application and the user applications in a Linera chain.

Re-exports

pub use system::SystemExecutionError;

pub use system::SystemExecutionStateView;

pub use system::SystemMessage;

pub use system::SystemOperation;

pub use system::SystemQuery;

pub use system::SystemResponse;

Modules

committee

system

test_utils

wasm_test

This assumes that the current directory is one of the crates.

Structs

ApplicationCallOutcome

The result of calling into a user application.

ApplicationRegistry

ApplicationRegistryView

Bytecode

A WebAssembly module’s bytecode.

BytecodeLocation

A reference to where the application bytecode is stored.

ChannelSubscription

The identifier of a channel, relative to a particular application.

ExecutionStateView

A view accessing the execution state of a chain.

FinalizeContext

InvalidWasmRuntime

Attempts to create an invalid WasmRuntime instance from a string.

MessageContext

OperationContext

QueryContext

RawExecutionOutcome

Externally visible results of an execution. These results are meant in the context of the application that created them.

RawOutgoingMessage

A message together with routing information.

ResourceControlPolicy

A collection of prices and limits associated with block execution.

ResourceController

ResourceTracker

The resources used so far by an execution process.

TestExecutionRuntimeContext

UserApplicationDescription

Description of the necessary information to run a user application.

Enums

ExecutionError

A type for errors happening during execution.

ExecutionOutcome

Externally visible results of an execution, tagged by their application.

ExecutionRuntimeConfig

System runtime implementation in use.

Message

A message to be sent and possibly executed in the receiver’s block.

MessageKind

The kind of outgoing message being sent.

Operation

An operation to be executed in a block.

Query

An query to be sent and possibly executed in the receiver’s block.

Response

The response to a query.

WasmContractModule

A user contract in a compiled WebAssembly module.

WasmExecutionError

Errors that can occur when executing a user application in a WebAssembly module.

WasmRuntime

The runtime to use for running the application.

WasmServiceModule

A user service in a compiled WebAssembly module.

Traits

BaseRuntime

ContractRuntime

ExecutionRuntimeContext

Requirements for the extra field in our state views (and notably the ExecutionStateView).

ServiceRuntime

UserContract

The public entry points provided by the contract part of an application.

UserContractModule

A factory trait to obtain a UserContract from a UserContractModule

UserService

The public entry points provided by the service part of an application.

UserServiceModule

A factory trait to obtain a UserService from a UserServiceModule

WithWasmDefault

Trait used to select a default WasmRuntime, if one is available.

Type Aliases

ContractSyncRuntime

ServiceSyncRuntime

UserApplicationId

Alias for linera_base::identifiers::ApplicationId. Use this alias in the core protocol where the distinction with the more general enum GenericApplicationId matters.

UserContractCode

An implementation of UserContractModule.

UserContractInstance

An implementation of UserContract.

UserServiceCode

An implementation of UserServiceModule.

UserServiceInstance

An implementation of UserService.