Crate linera_execution

This module manages the execution of the system application and the user applications in a Linera chain.

Re-exports

pub use crate::committee::Committee;

pub use crate::execution_state_actor::ExecutionRequest;

pub use crate::execution_state_actor::ExecutionStateActor;

pub use crate::system::SystemExecutionStateView;

pub use crate::system::SystemMessage;

pub use crate::system::SystemOperation;

pub use crate::system::SystemQuery;

pub use crate::system::SystemResponse;

Modules

committee

evm

Support for user applications compiled for the EVM

execution_state_actor

Handle requests from the synchronous execution thread of user applications.

system

test_utils

wasm_test

This assumes that the current directory is one of the crates.

Structs

BaseRuntimeApi

An implementation of the runtime API used to access the common behaviour and the view storage for both contracts and services.

BlobState

The state of a blob of binary data.

ContractEntrypoints

ContractRuntimeApi

An implementation of the system API made available to contracts.

ExecutionRuntimeConfig

Configuration options for the execution runtime available to applications.

ExecutionStateView

A view accessing the execution state of a chain.

FinalizeContext

InvalidWasmRuntime

Attempts to create an invalid WasmRuntime instance from a string.

JsVec

MessageContext

OperationContext

OutgoingMessage

A posted message together with routing information.

ProcessStreamsContext

QueryContext

QueryOutcome

The outcome of the execution of a query.

ResourceControlPolicy

A collection of prices and limits associated with block execution.

ResourceController

ResourceTracker

The resources used so far by an execution process. Acts as an accumulator for all resources consumed during a specific execution flow. This could be the execution of a block, the processing of a single message, or a specific phase within these broader operations.

RuntimeApiData

Common host data used as the UserData of the system API implementations.

ServiceEntrypoints

ServiceRuntimeApi

An implementation of the system API made available to services.

ServiceRuntimeEndpoint

How to interact with a long-lived service runtime.

ServiceSyncRuntime

TestExecutionRuntimeContext

ThreadPool

A pool of shared resources, each of which can only be used once at a time.

TransactionOutcome

The TransactionTracker contents after a transaction has finished.

TransactionTracker

Tracks oracle responses and execution outcomes of an ongoing transaction execution, as well as replayed oracle responses.

UserContractCode

An implementation of UserContractModule.

UserServiceCode

An implementation of UserServiceModule.

Enums

EvmRuntime

ExecutionError

A type for errors happening during execution.

Message

A message to be sent and possibly executed in the receiver’s block.

MessageKind

The kind of outgoing message being sent.

Operation

An operation to be executed in a block.

Query

An query to be sent and possibly executed in the receiver’s block.

QueryResponse

The response to a query.

ServiceRuntimeRequest

A request to the service runtime actor.

WasmContractModule

A user contract in a compiled WebAssembly module.

WasmExecutionError

Errors that can occur when executing a user application in a WebAssembly module.

WasmRuntime

The runtime to use for running the application.

WasmServiceModule

A user service in a compiled WebAssembly module.

Constants

LINERA_SOL

The Linera.sol library code to be included in solidity smart contracts using Linera features.

LINERA_TYPES_SOL

Traits

BalanceHolder

How to access the balance of an account.

BaseRuntime

ContractRuntime

ExecutionRuntimeContext

Requirements for the extra field in our state views (and notably the ExecutionStateView).

ServiceRuntime

UserContract

The public entry points provided by the contract part of an application.

UserContractModule

A factory trait to obtain a UserContract from a UserContractModule

UserService

The public entry points provided by the service part of an application.

UserServiceModule

A factory trait to obtain a UserService from a UserServiceModule

WithWasmDefault

Trait used to select a default WasmRuntime, if one is available.

Type Aliases

ContractSyncRuntimeHandle

ServiceSyncRuntimeHandle

UserContractInstance

An implementation of UserContract.

UserServiceInstance

An implementation of UserService.