Crate linera_execution

This module manages the execution of the system application and the user applications in a Linera chain.

Re-exports

pub use crate::system::SystemExecutionError;

pub use crate::system::SystemExecutionStateView;

pub use crate::system::SystemMessage;

pub use crate::system::SystemOperation;

pub use crate::system::SystemQuery;

pub use crate::system::SystemResponse;

Modules

committee

system

test_utils

wasm_test

This assumes that the current directory is one of the crates.

Structs

ApplicationCallOutcome

The result of calling into a user application.

ApplicationRegistry

ApplicationRegistryView

BlobState

The state of a blob of binary data.

ChannelSubscription

The identifier of a channel, relative to a particular application.

ContractEntrypoints

ContractSystemApi

An implementation of the system API made available to contracts.

ExecutionRuntimeConfig

Configuration options for the execution runtime available to applications.

ExecutionStateView

A view accessing the execution state of a chain.

FinalizeContext

InvalidWasmRuntime

Attempts to create an invalid WasmRuntime instance from a string.

MessageContext

OperationContext

QueryContext

RawExecutionOutcome

Externally visible results of an execution. These results are meant in the context of the application that created them.

RawOutgoingMessage

A message together with routing information.

ResourceControlPolicy

A collection of prices and limits associated with block execution.

ResourceController

ResourceTracker

The resources used so far by an execution process.

ServiceEntrypoints

ServiceRuntimeEndpoint

How to interact with a long-lived service runtime.

ServiceSyncRuntime

ServiceSystemApi

An implementation of the system API made available to services.

SystemApiData

Common host data used as the UserData of the system API implementations.

TestExecutionRuntimeContext

TransactionTracker

Tracks oracle responses and execution outcomes of an ongoing transaction execution, as well as replayed oracle responses.

UserContractCode

An implementation of UserContractModule.

UserServiceCode

An implementation of UserServiceModule.

ViewSystemApi

An implementation of the system API used to access the view storage for both contracts and services.

Enums

ExecutionError

A type for errors happening during execution.

ExecutionOutcome

Externally visible results of an execution, tagged by their application.

ExecutionRequest

Requests to the execution state.

Message

A message to be sent and possibly executed in the receiver’s block.

MessageKind

The kind of outgoing message being sent.

Operation

An operation to be executed in a block.

Query

An query to be sent and possibly executed in the receiver’s block.

Response

The response to a query.

ServiceRuntimeRequest

A request to the service runtime actor.

WasmContractModule

A user contract in a compiled WebAssembly module.

WasmExecutionError

Errors that can occur when executing a user application in a WebAssembly module.

WasmRuntime

The runtime to use for running the application.

WasmServiceModule

A user service in a compiled WebAssembly module.

Traits

BaseRuntime

ContractRuntime

ExecutionRuntimeContext

Requirements for the extra field in our state views (and notably the ExecutionStateView).

ServiceRuntime

UserContract

The public entry points provided by the contract part of an application.

UserContractModule

A factory trait to obtain a UserContract from a UserContractModule

UserService

The public entry points provided by the service part of an application.

UserServiceModule

A factory trait to obtain a UserService from a UserServiceModule

WithWasmDefault

Trait used to select a default WasmRuntime, if one is available.

Type Aliases

ContractSyncRuntimeHandle

ServiceSyncRuntimeHandle

UserContractInstance

An implementation of UserContract.

UserServiceInstance

An implementation of UserService.