linera_execution

Module system

Structs

CreateApplicationResult

EpochEventData

The data stored in an epoch creation event.

EventSubscriptions

The applications subscribing to a particular stream, and the next event index.

OpenChainConfig

The initial configuration for a new chain.

SystemExecutionStateView

A view accessing the execution state of the system of a chain.

SystemQuery

A query to the system state.

SystemResponse

The response to a system query.

UserData

Optional user message attached to a transfer.

Enums

AdminOperation

Operations that are only allowed on the admin chain.

SystemMessage

A system message meant to be executed on a remote chain.

SystemOperation

A system operation.

Statics

EPOCH_STREAM_NAME

The event stream name for new epochs and committees.

REMOVED_EPOCH_STREAM_NAME

The event stream name for removed epochs.