Module linera_execution::system

Structs

OpenChainConfig

The configuration for a new chain.

SystemExecutionStateView

A view accessing the execution state of the system of a chain.

SystemQuery

A query to the system state.

SystemResponse

The response to a system query.

UserData

Optional user message attached to a transfer.

Enums

AdminOperation

Operations that are only allowed on the admin chain.

Recipient

The recipient of a transfer.

SystemChannel

The channels available in the system application.

SystemExecutionError

SystemMessage

A system message meant to be executed on a remote chain.

SystemOperation

A system operation.

Statics

CREATE_APPLICATION_MESSAGE_INDEX

The relative index of the ApplicationCreated message created by the CreateApplication operation.

OPEN_CHAIN_MESSAGE_INDEX

The relative index of the OpenChain message created by the OpenChain operation.