linera_execution

Module test_utils

Modules

solidity

Code for compiling solidity smart contracts for testing purposes.

Structs

MockApplication

A mocked implementation of a user application.

MockApplicationInstance

A mocked implementation of a user application instance.

SystemExecutionState

A system execution state, not represented as a view but as a simple struct.

Enums

ExpectedCall

An expected call to a MockApplicationInstance.

Traits

RegisterMockApplication

Registration of MockApplications to use in tests.

Functions

blob_oracle_responses

Creates a vector of OracleResponses for the supplied BlobIds.

create_dummy_message_context

Creates a dummy MessageContext to use in tests.

create_dummy_operation_context

Creates a dummy OperationContext to use in tests.

create_dummy_query_context

Creates a dummy QueryContext to use in tests.

create_dummy_user_application_description

Creates a dummy ApplicationDescription for use in tests.

create_dummy_user_application_registrations

dummy_chain_description

dummy_chain_description_with_owner

dummy_chain_description_with_ownership_and_balance

dummy_committee

dummy_committees

test_accounts_strategy

Creates a Strategy for creating a BTreeMap of AccountOwners with an initial non-zero Amount of tokens.